Fix untuk bagian kriteria dan fix bug

+++ database/migrations/2023_04_17_122102_create_sub_kriteria_penilaians_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->uuid('id')->primary();
             $table->foreignUuid('id_kriteria')->constrained('kriteria_penilaians')->cascadeOnDelete();
             $table->string('nama_sub_kriteria');
-            $table->tinyInteger('nilai_sub_kriteria');
+            $table->float('nilai_sub_kriteria');
             $table->timestamps();
         });
     }
